requirements.txt
altair==6.1.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.29.0
Flask==3.1.3
fsspec==2026.4.0
gitdb==4.0.12
GitPython==3.1.50
h11==0.16.0
httptools==0.7.1
idna==3.14
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.20.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.2
mpmath==1.3.0
narwhals==2.21.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
packaging==26.2
pandas==3.0.2
parso==0.8.7
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==24.0.0
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.28
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
starlette==1.0.0
streamlit==1.57.0
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5.5
traitlets==5.15.0
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.46.0
wcwidth==0.7.0
websockets==16.0
Werkzeug==3.1.8
